Navigation push 2, Hover and animation would be done later
