+++ src/Assets/index.js
@@ -1,0 +1,20 @@
+import home from  "./home.svg";
+import file from  "./file.svg";
+import file2 from  "./file2.svg";
+import fileRequest from  "./fileRequest.svg";
+import share from  "./share.svg";
+import star from "./star.svg";
+import filesLogo from "./files-logo.png";
+import bin from "./bin.svg";
+
+
+export {
+        home,
+        file,
+        file2,
+        fileRequest,
+        share,
+        star,
+        filesLogo,
+        bin,
+    };
