+++ src/Pages/MyFiles.jsx
@@ -2,8 +2,9 @@ import React from 'react';
 
 const MyFiles = () => {
     return (
-        <div>
-            <h1>My Files</h1>
+        <div className="flex flex-col justify-center items-center h-screen font-lora text-7xl bg-primary">
+            <h1>My Files</h1> 
+            
             <p>Welcome to your files page!</p>
         </div>
     );
